codes/requirements.txt
torch>=1.10.0
torchvision>=0.11.1
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=0.24.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
tensorboard>=2.5.0
opencv-python>=4.5.0
Pillow>=8.3.0
tqdm>=4.60.0
pyyaml>=5.4.0
mne>=0.23.0
pywt>=1.1.0
h5py>=3.3.0
joblib>=1.0.0
black>=21.0
flake8>=3.9.0
pytest>=6.2.0